fix: Update vue-apollo.js template to clarify the link option (#62)
The link option in vue-apollo.js option is misleading, and suggests that you can pass in an httpLink object to override it. Behind the scenes, though, the links are concatted, which results in the console warning about concating terminating links. Other links compositions can be provided to the link option, but passing httpLink won't work and should be handled in `httpLinkOptions` instead. Wording is up to you, but this made me stumble more than it should have.